fix: support container names from specific docker registries
The executor-docker assumes that users will pull images from Docker Hub.
Unfortunately, this causes odd behavior with images that are hosted on
private docker registries, such as personal registries or otherwise. This
fix is to parse the image name by parts. Based on context and what is
given, it will try its best to build an image name and its corresponding
tag name.